remove base document, does not work for other types
